[chore][receiver/prometheusreceiver]: fix Openmetrics tests (#42417)
#### Description

The current test doesn't actually work most of the time as it only
asserts on having scraped series when there's a single set of resource
metrics in the result.

Most of the times scrape is so fast it does two scrapes, the second
being unsuccessful and having "up" == 0 and metrics stale in it.

I've fixed the assert function to always actually assert.

Consequences:

Some histogram test cases don't match the implementation where we ignore
such histograms that lack the Count field. Since the tests aren't about
missing Count, I've fixed them.

Some test cases are valid but don't have series in them, so those need a
different assert.

We could consider adding a test case for missing count.

#### Link to tracking issue
Fixes #42376

#### Testing
Unit tests.

#### Documentation
N/A

---------

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>