[receiver/libhoney] Set scopes more flexibly to avoid losing service.name (#42432)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

A bug happens when a library.name isn't set but service.name is set.
This causes the service.name to be dropped since the scope includes
both.

This change adds placeholder scopes for each service.name that comes in,
in the event that no library.name exists it has a fall-back.

This seems to be triggered by Jaeger traces that are converted into otlp
since they do not have instrumentation_scope name or version.