[exporter/datadog] Add 'exporter.datadogexporter.InferIntervalForDeltaMetrics' feature gate (#42494)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Adds a new `exporter.datadogexporter.InferIntervalForDeltaMetrics`
feature gate. This feature gate will set the interval field for OTLP
delta metrics when it can be inferred (i.e. when the difference between
Timestamp and StartTimestamp is close enough to a whole number of
seconds).

Relates to
[DataDog/opentelemetry-mapping-go/pull/765](https://github.com/DataDog/opentelemetry-mapping-go/pull/765)