receiver/postgresql - Include execution time metric into query samples (#42428)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
The logs collection in PostgreSQL receiver currently does not include
execution time of the currently running queries (query_samples) in
payload. This needs to be calculated in the collector query itself based
on the query_start time.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/42429

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Unit tests added/updated

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->